PUBLIC BETA, 40% complete
A really early version of the project that isnt going to look like the finished product, its used to make the connections and test the interactions between files. It has the login and register system and only a part of the main task part, to be added session settings, user profile and 6 other small changes.